requirements.txt
minecraft-launcher-lib
pillow
